+++ include/openPMD/Record.hpp
@@ -38,6 +38,7 @@ class Record : public BaseRecord< RecordComponent >
 
 public:
     Record(Record const&);
+    Record& operator=(Record const&);
     ~Record() override;
 
     Record& setUnitDimension(std::map< UnitDimension, double > const&);